ingreso de autenticacion

+++ routes/admin.php
@@ -6,6 +6,6 @@ use App\Http\Controllers\TicketController;
 use App\Models\Pago;
 use Illuminate\Support\Facades\Route;
 
-Route::get('/admin', [PagoController::class, 'index'])->name('pago.index');
-Route::delete('/admin/{id_pago}', [PagoController::class, 'destroy'])->name('pago.destroy');
-Route::get('/tickets/show', [TicketController::class, 'show'])->name('admin.showticket');
+Route::get('/admin', [PagoController::class, 'index'])->name('pago.index')->middleware('auth');
+Route::delete('/admin/{id_pago}', [PagoController::class, 'destroy'])->name('pago.destroy')->middleware('auth');
+Route::get('/tickets/show', [TicketController::class, 'show'])->name('admin.showticket')->middleware('auth');
